requirements.txt
ipython
keras
livelossplot
matplotlib
numpy
Pillow
seaborn
tensorflow
utils
